+++ rust/crd/src/druidconnection.rs
@@ -17,10 +17,14 @@ type Result<T, E = Error> = std::result::Result<T, E>;
 #[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct ClusterRef {
+    /// The name of the stacklet.
     pub name: String,
+    /// The namespace. Defaults to the namespace of the `DruidConnection` if it is not specified.
     pub namespace: Option<String>,
 }
 
+/// The DruidConnection resource can be used to automatically deploy a Druid datasource in Superset.
+/// Learn more about it in the [Superset operator usage guide](DOCS_BASE_URL_PLACEHOLDER/superset/usage-guide/connecting-druid).
 #[derive(Clone, CustomResource, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
 #[kube(
     group = "superset.stackable.tech",
@@ -37,7 +41,9 @@ pub struct ClusterRef {
 )]
 #[serde(rename_all = "camelCase")]
 pub struct DruidConnectionSpec {
+    /// The Superset to connect.
     pub superset: ClusterRef,
+    /// The Druid to connect.
     pub druid: ClusterRef,
 }